abstracted from the cost function used in softmarginloss
The soft margin loss is defined as

  l(x,y,w) = max_y' Γ(y,y')<w,Θ(x,y')-Θ(x,y)> + Δ(y,y')

where Δ(y,y') is the cost function. This function can now be abitrary,
privided it is linear in y', i.e.,

  Δ(y,y') = c(y) + <a(y),y'>.

For any linear cost function, it is therefore sufficient to provide the
constant c(y) and linear a(y) contribution to the loss function. This is
reflected in the abstract class Cost now.

The Hamming cost is now just an instantiation of this abstract cost
function class with

  c(y) = <y,1> and a(y) = 1-2y

where "1" denotes a vector of 1s.

The cost function used by the soft margin loss can now be set via the
configuration string

  Loss.costFunctionType = <costfunction>

where <costfunction> can currently only be HAMMING.